refactor(backend): move compute_next_reset to domain.dates layer
An ORM model (models/user.py) imported compute_next_reset from services/usage,
inverting the models -> services -> routers direction. The helper is pure
calendar arithmetic whose only dependency is domain.dates.ensure_aware, so move
it into domain/dates.py (its natural home) and re-export from services/usage to
keep existing callers working. models/user.py now imports from domain.dates.
Its unit tests move to tests/domain/test_dates.py alongside the code; behaviour
is unchanged.
